Move analysis classes
