+++ core/src/main/java/org/eqasim/core/analysis/legs/LegItem.java
@@ -1,4 +1,4 @@
-package org.eqasim.core.analysis;
+package org.eqasim.core.analysis.legs;
 
 import org.matsim.api.core.v01.Coord;
 import org.matsim.api.core.v01.Id;
