+++ core/src/main/java/org/eqasim/core/analysis/legs/LegWriter.java
@@ -1,4 +1,4 @@
-package org.eqasim.core.analysis;
+package org.eqasim.core.analysis.legs;
 
 import java.io.BufferedWriter;
 import java.io.FileOutputStream;
@@ -6,19 +6,21 @@ import java.io.IOException;
 import java.io.OutputStreamWriter;
 import java.util.Collection;
 
-public class TripWriter {
-	final private Collection<TripItem> trips;
+import org.eqasim.core.analysis.DistanceUnit;
+
+public class LegWriter {
+	final private Collection<LegItem> legs;
 	final private String delimiter;
 
 	final private DistanceUnit inputUnit;
 	final private DistanceUnit outputUnit;
 
-	public TripWriter(Collection<TripItem> trips, DistanceUnit inputUnit, DistanceUnit outputUnit) {
-		this(trips, inputUnit, outputUnit, ";");
+	public LegWriter(Collection<LegItem> legs, DistanceUnit inputUnit, DistanceUnit outputUnit) {
+		this(legs, inputUnit, outputUnit, ";");
 	}
 
-	public TripWriter(Collection<TripItem> trips, DistanceUnit inputUnit, DistanceUnit outputUnit, String delimiter) {
-		this.trips = trips;
+	public LegWriter(Collection<LegItem> legs, DistanceUnit inputUnit, DistanceUnit outputUnit, String delimiter) {
+		this.legs = legs;
 		this.delimiter = delimiter;
 		this.inputUnit = inputUnit;
 		this.outputUnit = outputUnit;
@@ -30,8 +32,8 @@ public class TripWriter {
 		writer.write(formatHeader() + "\n");
 		writer.flush();
 
-		for (TripItem trip : trips) {
-			writer.write(formatTrip(trip) + "\n");
+		for (LegItem leg : legs) {
+			writer.write(formatLeg(leg) + "\n");
 			writer.flush();
 		}
 
@@ -39,14 +41,11 @@ public class TripWriter {
 		writer.close();
 	}
 
-	private String normalizeActivityType(String activityType) {
-		return activityType.replaceAll("_[0-9]+$", "");
-	}
-
 	private String formatHeader() {
 		return String.join(delimiter, new String[] { //
 				"person_id", //
 				"person_trip_id", //
+				"leg_index", //
 				"origin_x", //
 				"origin_y", //
 				"destination_x", //
@@ -55,9 +54,6 @@ public class TripWriter {
 				"vehicle_distance", //
 				"routed_distance", //
 				"mode", //
-				"preceding_purpose", //
-				"following_purpose", //
-				"returning", //
 				"euclidean_distance" //
 		});
 	}
@@ -88,26 +84,24 @@ public class TripWriter {
 		return factor;
 	}
 
-	private String formatTrip(TripItem trip) {
+	private String formatLeg(LegItem leg) {
 		double inputFactor = getUnitFactor(inputUnit);
 		double outputFactor = 1.0 / getUnitFactor(outputUnit);
 
 		return String.join(delimiter, new String[] { //
-				trip.personId.toString(), //
-				String.valueOf(trip.personTripId), //
-				String.valueOf(trip.origin.getX()), //
-				String.valueOf(trip.origin.getY()), //
-				String.valueOf(trip.destination.getX()), //
-				String.valueOf(trip.destination.getY()), //
-				String.valueOf(trip.departureTime), //
-				String.valueOf(trip.travelTime), //
-				String.valueOf(trip.vehicleDistance * inputFactor * outputFactor), //
-				String.valueOf(trip.routedDistance * inputFactor * outputFactor), //
-				String.valueOf(trip.mode), //
-				normalizeActivityType(String.valueOf(trip.precedingPurpose)), //
-				normalizeActivityType(String.valueOf(trip.followingPurpose)), //
-				String.valueOf(trip.returning), //
-				String.valueOf(trip.euclideanDistance * inputFactor * outputFactor) //
+				leg.personId.toString(), //
+				String.valueOf(leg.personTripId), //
+				String.valueOf(leg.legIndex), //
+				String.valueOf(leg.origin.getX()), //
+				String.valueOf(leg.origin.getY()), //
+				String.valueOf(leg.destination.getX()), //
+				String.valueOf(leg.destination.getY()), //
+				String.valueOf(leg.departureTime), //
+				String.valueOf(leg.travelTime), //
+				String.valueOf(leg.vehicleDistance * inputFactor * outputFactor), //
+				String.valueOf(leg.routedDistance * inputFactor * outputFactor), //
+				String.valueOf(leg.mode), //
+				String.valueOf(leg.euclideanDistance * inputFactor * outputFactor) //
 		});
 	}
 }
