+++ core/src/main/java/org/eqasim/core/analysis/trips/RunTripAnalysis.java
@@ -1,4 +1,4 @@
-package org.eqasim.core.analysis;
+package org.eqasim.core.analysis.trips;
 
 import java.io.IOException;
 import java.util.Arrays;
@@ -7,6 +7,10 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.apache.log4j.Logger;
+import org.eqasim.core.analysis.DefaultPersonAnalysisFilter;
+import org.eqasim.core.analysis.DistanceUnit;
+import org.eqasim.core.analysis.PersonAnalysisFilter;
+import org.eqasim.core.components.EqasimMainModeIdentifier;
 import org.matsim.api.core.v01.Scenario;
 import org.matsim.api.core.v01.network.Network;
 import org.matsim.core.config.CommandLine;
@@ -15,12 +19,13 @@ import org.matsim.core.config.Config;
 import org.matsim.core.config.ConfigUtils;
 import org.matsim.core.network.NetworkUtils;
 import org.matsim.core.network.io.MatsimNetworkReader;
+import org.matsim.core.router.MainModeIdentifier;
 import org.matsim.core.scenario.ScenarioUtils;
 import org.matsim.facilities.ActivityFacilities;
 import org.matsim.facilities.MatsimFacilitiesReader;
 
-public class RunLegAnalysis {
-	private final static Logger logger = Logger.getLogger(RunLegAnalysis.class);
+public class RunTripAnalysis {
+	private final static Logger logger = Logger.getLogger(RunTripAnalysis.class);
 
 	static public void main(String[] args) throws IOException, ConfigurationException {
 		CommandLine cmd = new CommandLine.Builder(args) //
@@ -55,10 +60,12 @@ public class RunLegAnalysis {
 
 		String outputPath = cmd.getOptionStrict("output-path");
 
+		MainModeIdentifier mainModeIdentifier = new EqasimMainModeIdentifier();
+
 		Collection<String> vehicleModes = Arrays.asList(cmd.getOption("vehicle-modes").orElse("car,pt").split(","))
 				.stream().map(s -> s.trim()).collect(Collectors.toSet());
 
-		Collection<LegItem> legs = null;
+		Collection<TripItem> trips = null;
 
 		if (cmd.hasOption("events-path")) {
 			String networkPath = cmd.getOptionStrict("network-path");
@@ -66,8 +73,8 @@ public class RunLegAnalysis {
 			new MatsimNetworkReader(network).readFile(networkPath);
 
 			String eventsPath = cmd.getOptionStrict("events-path");
-			LegListener legListener = new LegListener(network, personAnalysisFilter);
-			legs = new LegReaderFromEvents(legListener).readLegs(eventsPath);
+			TripListener tripListener = new TripListener(network, mainModeIdentifier, personAnalysisFilter);
+			trips = new TripReaderFromEvents(tripListener).readTrips(eventsPath);
 		} else {
 			Network network = null;
 			ActivityFacilities facilities = null;
@@ -89,13 +96,13 @@ public class RunLegAnalysis {
 			}
 
 			String populationPath = cmd.getOptionStrict("population-path");
-			legs = new LegReaderFromPopulation(vehicleModes, personAnalysisFilter, Optional.ofNullable(network),
-					Optional.ofNullable(facilities)).readLegs(populationPath);
+			trips = new TripReaderFromPopulation(vehicleModes, mainModeIdentifier, personAnalysisFilter,
+					Optional.ofNullable(network), Optional.ofNullable(facilities)).readTrips(populationPath);
 		}
 
 		DistanceUnit inputUnit = DistanceUnit.valueOf(cmd.getOption("input-distance-unit").orElse("meter"));
 		DistanceUnit outputUnit = DistanceUnit.valueOf(cmd.getOption("output-distance-unit").orElse("meter"));
 
-		new LegWriter(legs, inputUnit, outputUnit).write(outputPath);
+		new TripWriter(trips, inputUnit, outputUnit).write(outputPath);
 	}
 }
