+++ core/src/main/java/org/eqasim/core/analysis/trips/TripListenerItem.java
@@ -1,4 +1,4 @@
-package org.eqasim.core.analysis;
+package org.eqasim.core.analysis.trips;
 
 import java.util.LinkedList;
 import java.util.List;
