+++ core/src/main/java/org/eqasim/core/analysis/trips/TripReaderFromPopulation.java
@@ -1,10 +1,11 @@
-package org.eqasim.core.analysis;
+package org.eqasim.core.analysis.trips;
 
 import java.util.Collection;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
 
+import org.eqasim.core.analysis.PersonAnalysisFilter;
 import org.matsim.api.core.v01.Coord;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.Scenario;
