+++ core/src/main/java/org/eqasim/core/simulation/analysis/EqasimAnalysisModule.java
@@ -1,10 +1,10 @@
 package org.eqasim.core.simulation.analysis;
 
 import org.eqasim.core.analysis.DefaultPersonAnalysisFilter;
-import org.eqasim.core.analysis.LegListener;
 import org.eqasim.core.analysis.PersonAnalysisFilter;
-import org.eqasim.core.analysis.TripListener;
+import org.eqasim.core.analysis.legs.LegListener;
 import org.eqasim.core.analysis.pt.PublicTransportLegListener;
+import org.eqasim.core.analysis.trips.TripListener;
 import org.matsim.api.core.v01.network.Network;
 import org.matsim.core.controler.AbstractModule;
 import org.matsim.core.router.MainModeIdentifier;
